Renaming the ProxyModel to OpaqueKeyProxyModel
